Move modal test infrastructure from mng to mng_modal
Move modal acceptance/release tests out of libs/mng into libs/mng_modal:
- test_exec.py (all modal) -> mng_modal/test_exec.py
- test_provision.py modal test -> mng_modal/test_provision.py
- test_snapshot.py modal tests -> mng_modal/test_snapshot.py

Remove modal fixtures from mng's conftest:
- modal_subprocess_env and session-scoped dependencies
- Modal token loading from setup_test_mng_env
- Modal-only imports (ModalSubprocessTestEnv, generate_test_environment_name, etc.)
- Modal marker registration from project-level conftest

Add modal subprocess fixtures to mng_modal's conftest:
- modal_subprocess_env and session-scoped dependencies
- setup_test_mng_env override with modal token loading

Add modal subprocess fixtures to mng_claude's conftest:
- Same fixtures so test_claude_agent_modal.py can access modal_subprocess_env
- setup_test_mng_env override with modal token loading
