+++ libs/mng/conftest.py
@@ -23,7 +23,6 @@ register_marker("rsync: marks tests that invoke rsync for file transfer")
 register_marker("unison: marks tests that start a real unison file-sync process")
 register_marker("docker: marks tests that invoke the docker CLI via subprocess")
 register_marker("docker_sdk: marks tests that use the Docker Python SDK in-process")
-register_marker("modal: marks tests that connect to the Modal cloud service")
 register_resource_guard("tmux")
 register_resource_guard("rsync")
 register_resource_guard("unison")
